+++ Verticality/Moves/Climb/Grab.cs
@@ -1,78 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
-using System.Linq;
-using System.Numerics;
-using System.Runtime.CompilerServices;
-using System.Security.Cryptography;
-using System.Text;
-using System.Threading.Tasks;
-using Vintagestory.API.Client;
 using Vintagestory.API.Common;
-using Vintagestory.API.Common.Entities;
 using Vintagestory.API.MathTools;
-using Vintagestory.API.Util;
-using Vintagestory.Common;
-using Vintagestory.Server;
-using static Verticality.CollisionUtils;
+using static Verticality.Lib.CollisionUtils;
 
-namespace Verticality
+namespace Verticality.Moves.Climb
 {
-    internal class EntityBehaviorClimb : EntityBehavior
-    {
-        public const float minHeight = 0.5f; // minimum height from player's feet
-        public const float maxHeight = 0.7f; // maximum height from player's LocalEyePos
-        public const float grabDistance = 0.5f; // maximum distance to grab point
-
-        private Grab grab;
-
-        public bool ClimbKeyDown
-        {
-            get
-            {
-                return ((ICoreClientAPI)entity.Api).Input.IsHotKeyPressed("climb");
-            }
-        }
-
-        public EntityBehaviorClimb(Entity entity) : base(entity) {}
-
-        public override string PropertyName()
-        {
-            return "climb";
-        }
-
-        public override void OnGameTick(float deltaTime)
-        {
-            base.OnGameTick(deltaTime);
-
-            if (entity.World.Side != EnumAppSide.Client) return;
-
-            EntityPlayer player = (EntityPlayer)entity;
-
-            if (ClimbKeyDown)
-            {
-                if (grab == null)
-                {
-                    grab = Grab.TryGrab(player);
-                } else
-                {
-                    if (grab.CanStillGrab())
-                    {
-                        player.Properties.CanClimbAnywhere = true;
-                    } else
-                    {
-                        grab = Grab.TryGrab(player);
-                        if (grab == null) player.Properties.CanClimbAnywhere = false;
-                    }
-                }
-            } else if (grab != null)
-            {
-                player.Properties.CanClimbAnywhere = false;
-                grab = null;
-            }
-        }
-    }
-
     internal class Grab
     {
         static float minHeight => EntityBehaviorClimb.minHeight;
@@ -140,7 +73,7 @@ namespace Verticality
             float yaw = entity.BodyYaw - GameMath.PIHALF;
 
             List<BlockPos> posList = new List<BlockPos>();
-            
+
             /*
             debugParticles.Color = ColorUtil.ColorFromRgba(255, 0, 0, 255);
             //*/
@@ -186,7 +119,7 @@ namespace Verticality
             }
 
 
-            Vec3d collPos = GetClosestPoint(collBoxes, entity.Pos.XYZ.AddCopy(0,minHeight,0));
+            Vec3d collPos = GetClosestPoint(collBoxes, entity.Pos.XYZ.AddCopy(0, minHeight, 0));
 
             Vec3d topPos = ToTheTop(collBoxes, collPos);
