+++ Verticality/changes.txt
@@ -1,3 +1,3 @@
 - removed the config file entirely
 - maximum height is now calculated from player's eye height, rather than feet height
-- added a charged jump
+- new move: charged jump
